pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "streamlit>=1.54.0",
]
